chore: migrate to pnpm workspace
Convert npm workspaces to pnpm (pnpm-workspace.yaml, workspace:* internal deps); packageManager pnpm@9.15.0; declare phantom deps (drizzle-orm, zod, vite); convert npm commands in source/tests/docs/CI to pnpm. Fix pre-existing breakage surfaced by verification: time-relative the date-rot extension-sync test, move the stray search export out of the +server.ts route so the web build passes, eslint-ignore the vitepress cache, add .prettierignore and format. Gitignore .claude/ and daemon/tmp/.
