+++ .gitignore
@@ -54,3 +54,5 @@ web/.svelte-kit/
 web/build/
 .claude/settings.local.json
 shared/src/platforms/reddit/cache/
+.claude/
+daemon/tmp/
