+++ README.md
@@ -14,22 +14,22 @@ cd pitchbox
 cp .env.example .env
 # edit .env - set DATABASE_URL, PITCHBOX_ROOT (absolute path), ENCRYPTION_KEY (openssl rand -hex 32)
 docker compose up -d postgres
-npm install
-npm run migrate
-npm run -w @pitchbox/shared seed:core
-npm run dev            # dashboard at http://127.0.0.1:5180
+pnpm install
+pnpm run migrate
+pnpm -F @pitchbox/shared seed:core
+pnpm run dev            # dashboard at http://127.0.0.1:5180
 ```
 
-Embedded daemon mode runs the scheduler + reply poller + retention loops inside the web process - set `PITCHBOX_EMBED_DAEMON=1` in `.env` and skip `npm run -w daemon dev`. See [the daemon docs](https://fiorelorenzo.github.io/pitchbox/daemon) for when to run it as a separate process instead.
+Embedded daemon mode runs the scheduler + reply poller + retention loops inside the web process - set `PITCHBOX_EMBED_DAEMON=1` in `.env` and skip `pnpm -F daemon dev`. See [the daemon docs](https://fiorelorenzo.github.io/pitchbox/daemon) for when to run it as a separate process instead.
 
-Prerequisites: Node >= 22, Docker, and one of the supported ACP-compatible agent CLIs on PATH (Claude Code, Codex, Gemini CLI, GitHub Copilot CLI, opencode, or Qwen Code - details in [docs/runners.md](./docs/runners.md)).
+Prerequisites: Node >= 22, pnpm 9.15 (via corepack), Docker, and one of the supported ACP-compatible agent CLIs on PATH (Claude Code, Codex, Gemini CLI, GitHub Copilot CLI, opencode, or Qwen Code - details in [docs/runners.md](./docs/runners.md)).
 
 ## Browser extension
 
 The companion Chrome extension auto-marks drafts as sent when you submit on Reddit, syncs replies back into the dashboard, and surfaces what it's doing in a persistent **side panel** with three tabs (Dashboard, Activity log, Settings).
 
 ```bash
-npm run build:extension   # then load extension/dist/ unpacked in chrome://extensions
+pnpm run build:extension   # then load extension/dist/ unpacked in chrome://extensions
 ```
 
 Click the toolbar icon to open the side panel. Pair with your dashboard in one click, kick off a manual sync, follow every operation in real time (pairings, DM/chat sync runs, Matrix token captures, Reddit actions), and tune theme (light/dark/system), language (en/it), density and poller interval from Settings. You can pair multiple backends (cloud + self-hosted) at the same time. Details: <https://fiorelorenzo.github.io/pitchbox/extension>
