+++ cli/package.json
@@ -13,8 +13,9 @@
     "typecheck": "tsc --noEmit"
   },
   "dependencies": {
-    "@pitchbox/shared": "*",
+    "@pitchbox/shared": "workspace:*",
     "commander": "^12.1.0",
+    "drizzle-orm": "^0.45.2",
     "zod": "^3.23.0",
     "dotenv": "^16.4.0"
   },
