+++ cli/tests/commands/skill-generate-start.test.ts
@@ -4,7 +4,7 @@ import { getDb, getPool, schema } from '@pitchbox/shared/db';
 import { eq, sql } from 'drizzle-orm';
 
 function cli(args: string): string {
-  return execSync(`npm -s run -w @pitchbox/cli dev -- ${args}`, {
+  return execSync(`pnpm -s -F @pitchbox/cli dev ${args}`, {
     encoding: 'utf8',
     cwd: '/Users/lorenzofiore/Progetti/Personale/pitchbox',
   });
