+++ docs/self-hosting.md
@@ -34,7 +34,7 @@ The daemon prunes ageing event logs and terminal drafts on a configurable schedu
 
 ## Upgrades
 
-`git pull && npm install && npm run migrate && npm run -w @pitchbox/shared seed:core`. The seed step refreshes built-in playbooks but leaves user-created rows alone.
+`git pull && pnpm install && pnpm run migrate && pnpm -F @pitchbox/shared seed:core`. The seed step refreshes built-in playbooks but leaves user-created rows alone.
 
 ## Dependency pinning policy
 
@@ -44,7 +44,7 @@ A few dependencies are pinned to **exact** versions (no caret) in `package.json`
 - Packages that ship native bindings (e.g. `sharp`) - minor bumps frequently change the prebuilt-binary matrix and break Docker images on uncommon architectures.
 - Build tooling whose output is shipped to users (`vite`, `vitest`, `@crxjs/vite-plugin`) - patch releases here can change the bundle layout or extension manifest in subtle ways.
 
-When adding or upgrading any of the above, write the exact version (no `^`, no `~`) and regenerate the lockfile with `npm install --package-lock-only`. Other dependencies may continue to use `^` ranges.
+When adding or upgrading any of the above, write the exact version (no `^`, no `~`) and regenerate the lockfile with `pnpm install --lockfile-only`. Other dependencies may continue to use `^` ranges.
 
 ## Logs
