+++ extension/src/lib/activity.ts
@@ -33,9 +33,7 @@ async function write(next: ActivityEvent[]): Promise<void> {
   await chrome.storage.local.set({ [KEY]: next });
 }
 
-export async function logEvent(
-  input: Omit<ActivityEvent, 'id' | 'ts'>,
-): Promise<ActivityEvent> {
+export async function logEvent(input: Omit<ActivityEvent, 'id' | 'ts'>): Promise<ActivityEvent> {
   const ev: ActivityEvent = { id: ulid(), ts: new Date().toISOString(), ...input };
   const current = await read();
   // Newest first; trim from the tail (oldest) when exceeding cap.
