+++ extension/src/sidepanel/app.css
@@ -97,7 +97,13 @@ html,
 body {
   background: var(--background);
   color: var(--foreground);
-  font-family: 'Inter Variable', -apple-system, BlinkMacSystemFont, 'Segoe UI', Helvetica, Arial,
+  font-family:
+    'Inter Variable',
+    -apple-system,
+    BlinkMacSystemFont,
+    'Segoe UI',
+    Helvetica,
+    Arial,
     sans-serif;
   margin: 0;
 }
