+++ package.json
@@ -4,36 +4,30 @@
   "private": true,
   "license": "AGPL-3.0-or-later",
   "type": "module",
-  "workspaces": [
-    "shared",
-    "cli",
-    "web",
-    "daemon",
-    "extension"
-  ],
+  "packageManager": "pnpm@9.15.0",
   "engines": {
     "node": ">=22"
   },
   "scripts": {
-    "build": "npm run build --workspaces --if-present",
+    "build": "pnpm -r --if-present run build",
     "test": "vitest run",
     "test:watch": "vitest",
     "lint": "eslint . && prettier --check .",
     "format": "prettier --write .",
-    "typecheck": "tsc --noEmit -p shared && tsc --noEmit -p cli && tsc --noEmit -p daemon && npm run -w @pitchbox/extension check && npm run -w web check",
-    "migrate": "npm run -w shared migrate",
-    "migrate:generate": "npm run -w shared migrate:generate",
+    "typecheck": "tsc --noEmit -p shared && tsc --noEmit -p cli && tsc --noEmit -p daemon && pnpm -F @pitchbox/extension check && pnpm -F web check",
+    "migrate": "pnpm -F shared migrate",
+    "migrate:generate": "pnpm -F shared migrate:generate",
     "db:up": "docker compose up -d postgres",
     "db:down": "docker compose down",
-    "dev": "concurrently -n web,ext,docs -c blue,magenta,green --kill-others-on-fail \"npm run -w web dev\" \"npm run -w @pitchbox/extension dev\" \"npm run docs:dev\"",
-    "dev:web": "npm run -w web dev",
-    "dev:extension": "npm run -w @pitchbox/extension dev",
-    "dev:docs": "npm run docs:dev",
+    "dev": "concurrently -n web,ext,docs -c blue,magenta,green --kill-others-on-fail \"pnpm -F web dev\" \"pnpm -F @pitchbox/extension dev\" \"pnpm run docs:dev\"",
+    "dev:web": "pnpm -F web dev",
+    "dev:extension": "pnpm -F @pitchbox/extension dev",
+    "dev:docs": "pnpm run docs:dev",
     "migrate:from-ai-game": "tsx scripts/migrate-from-ai-game.ts",
     "seed:embertold": "tsx scripts/seed-embertold.ts",
-    "reseed": "npm run -w @pitchbox/shared seed:core && npm run seed:embertold",
-    "build:extension": "npm run -w @pitchbox/extension build",
-    "zip:extension": "npm run -w @pitchbox/extension zip",
+    "reseed": "pnpm -F @pitchbox/shared seed:core && pnpm run seed:embertold",
+    "build:extension": "pnpm -F @pitchbox/extension build",
+    "zip:extension": "pnpm -F @pitchbox/extension zip",
     "docs:dev": "vitepress dev docs --port 5181",
     "docs:build": "vitepress build docs",
     "docs:preview": "vitepress preview docs",
@@ -51,6 +45,7 @@
     "sharp": "0.34.5",
     "tsx": "^4.21.0",
     "typescript": "^5.9.3",
+    "vite": "^8.0.0",
     "vitepress": "^1.6.4",
     "vitest": "4.1.5"
   },
