+++ web/package.json
@@ -14,17 +14,19 @@
   },
   "dependencies": {
     "@bytemd/plugin-gfm": "^1.22.0",
-    "@pitchbox/daemon": "*",
-    "@pitchbox/shared": "*",
+    "@pitchbox/daemon": "workspace:*",
+    "@pitchbox/shared": "workspace:*",
     "@tailwindcss/vite": "^4.2.4",
     "@types/dompurify": "^3.0.5",
     "bits-ui": "^2.18.0",
     "bytemd": "^1.22.0",
     "diff": "^5.2.2",
     "dompurify": "^3.4.1",
+    "drizzle-orm": "^0.45.2",
     "lucide-svelte": "^1.0.1",
     "marked": "^18.0.2",
-    "tailwindcss": "^4.2.4"
+    "tailwindcss": "^4.2.4",
+    "zod": "^3.23.0"
   },
   "devDependencies": {
     "@fontsource-variable/inter": "^5.2.8",
